26/02/06 정재은 등록자 수정,..등등

+++ src/main/java/com/yedam/app/attach/service/AttachmentVO.java
@@ -9,7 +9,7 @@ import lombok.Data;
 @Data
 public class AttachmentVO {
 	private Long fileCode;
-	private Long userCode;
+	private Integer userCode;
 	private String tableCode;
 
 	@DateTimeFormat(pattern = "yyyy-MM-dd'T'HH:mm")
